Task 2
Visual Studio Code and Golang were chosen as the primary technologies that will be utilized. A project folder was created in order to contain the files main.go and file1.txt. At the moment, the main.go file contains code which prompts the user for a file name and then reads the file. The program is also able to check for reading errors, and it outputs all the contents of the file.